+++ history.md
@@ -1,6 +1,6 @@
 # 2024-08-12
 
-- Added `coc.preferences.formatterExtension` extension
+- Added `coc.preferences.formatterExtension` configuration
 
 # 2024-07-04