+++ history.md
@@ -1,3 +1,7 @@
+# 2024-10-25
+
+- Mention [davidosomething/coc-diagnostics-shim.nvim](https://github.com/davidosomething/coc-diagnostics-shim.nvim) as alternative to ALE for diagnostics display.
+
 # 2024-08-28
 
 - Add configuration `codeLens.display`
@@ -26,7 +30,7 @@
 # 2024-05-29
 
 - Break change: increase minimum vim/nvim version requirement
-  - vim  9.0.0438
+  - vim 9.0.0438
   - nvim 0.8.0
 
 # 2024-05-14